Add abstract class to avoid duplications
